requirements.txt

google-adk
google-genai